add socket connection handler to dispatcher
