+++ Sources/Relayer/Dispatching.swift
@@ -16,24 +16,23 @@ final class Dispatcher: NSObject, Dispatching {
     var onMessage: ((String) -> ())?
     private var textFramesQueue = Queue<String>()
     private var networkMonitor: NetworkMonitoring
-    private let url: URL
     var socket: WebSocketSessionProtocol
     var socketConnectionObserver: SocketConnectionObserving
+    var socketConnectionHandler: SocketConnectionHandler
     
-    init(url: URL,
-         networkMonitor: NetworkMonitoring = NetworkMonitor(),
+    init(networkMonitor: NetworkMonitoring = NetworkMonitor(),
          socket: WebSocketSessionProtocol,
          socketConnectionObserver: SocketConnectionObserving,
          socketConnectionHandler: SocketConnectionHandler) {
-        self.url = url
         self.networkMonitor = networkMonitor
         self.socket = socket
         self.socketConnectionObserver = socketConnectionObserver
+        self.socketConnectionHandler = socketConnectionHandler
         super.init()
         setUpWebSocketSession()
         setUpSocketConnectionObserving()
         setUpNetworkMonitoring()
-        socket.connect(on: url)
+        socket.connect()
     }
 
     func send(_ string: String, completion: @escaping (Error?) -> Void) {
@@ -46,13 +45,13 @@ final class Dispatcher: NSObject, Dispatching {
     }
     
     func connect() {
-        if !socket.isConnected {
-            socket.connect(on: url)
-        }
+        //todo handle error
+        try! socketConnectionHandler.handleConnect()
     }
     
     func disconnect(closeCode: URLSessionWebSocketTask.CloseCode) {
-        socket.disconnect(with: closeCode)
+        //todo handle error
+        try! socketConnectionHandler.handleDisconnect()
     }
     
     private func setUpWebSocketSession() {
@@ -76,10 +75,10 @@ final class Dispatcher: NSObject, Dispatching {
     
     private func setUpNetworkMonitoring() {
         networkMonitor.onSatisfied = { [weak self] in
-            self?.connect()
+            self?.socketConnectionHandler.handleNetworkSatisfied()
         }
         networkMonitor.onUnsatisfied = { [weak self] in
-            self?.disconnect(closeCode: .goingAway)
+            self?.socketConnectionHandler.handleNetworkUnsatisfied()
         }
         networkMonitor.startMonitoring()
     }
@@ -96,73 +95,3 @@ final class Dispatcher: NSObject, Dispatching {
 }
 
 
-protocol SocketConnectionHandler {
-    var appStateObserver: AppStateObserving {get}
-    func handleConnect() throws
-    func handleDisconnect() throws
-    func handleNetworkUnsatisfied() throws
-    func handleNetworkSatisfied(_ url: URL)
-}
-
-
-import UIKit
-class AutomaticSocketConnectionHandler: SocketConnectionHandler {
-    enum Error: Swift.Error {
-        case manualSocketConnectionForbidden
-        case manualSocketDisconnectionForbidden
-    }
-    var appStateObserver: AppStateObserving
-    let socket: WebSocketSessionProtocol
-    let url: URL
-    private var backgroundTaskID: UIBackgroundTaskIdentifier = .invalid
-    
-    init(socket: WebSocketSessionProtocol, url: URL, appStateObserver: AppStateObserving = AppStateObserver()) {
-        self.appStateObserver = appStateObserver
-        self.socket = socket
-        self.url = url
-        setUpStateObserving()
-    }
-    
-    private func setUpStateObserving() {
-        appStateObserver.onWillEnterBackground = {
-            
-        }
-        appStateObserver.onWillEnterForeground = { [unowned self] in
-            socket.connect(on: url)
-        }
-    }
-    
-    func registerBackgroundTask() {
-#if os(iOS)
-        backgroundTaskID = UIApplication.shared.beginBackgroundTask (withName: "Finish Network Tasks") { [weak self] in
-            self?.endBackgroundTask()
-        }
-#endif
-    }
-    
-    func endBackgroundTask() {
-#if os(iOS)
-        socket.disconnect(with: .normalClosure)
-        UIApplication.shared.endBackgroundTask(backgroundTaskID)
-        backgroundTaskID = .invalid
-#endif
-    }
-    
-    func handleConnect() throws {
-        throw Error.manualSocketConnectionForbidden
-    }
-    
-    func handleDisconnect() throws {
-        throw Error.manualSocketDisconnectionForbidden
-    }
-    
-    func handleNetworkUnsatisfied() {
-        socket.disconnect(with: .goingAway)
-    }
-    
-    func handleNetworkSatisfied(_ url: URL) {
-        if !socket.isConnected {
-            socket.connect(on: url)
-        }
-    }
-}
